requirements.txt
requests==2.31.0
PyYAML==6.0.1
openai==1.30.1
PyGithub==2.3.0
flask==3.0.3
python-dotenv==1.0.1
lxml==5.2.1
packaging==24.0
GitPython==3.1.43
